configure: Support disabling specific languages [PR12407]
Sometimes it can be desirable to get the semantics of
--enable-languages=all, but to exclude one or more languages from the
build. Currently this is not directly supported; the best you can do is to
list the ones you do want to be built as arguments to --enable-languages.
In addition to being inconvenient, this also complicates cross-platform
portability, since --enable-languages=all carries the useful semantics that
unsupported languages will be skipped automatically; by contrast, languages
listed explicitly as arguments to --enable-languages will produce a hard
error if they are not supported.

This patch extends the syntax of --enable-languages so that, e.g.:

--enable-languages=all,^xyz,^abc

would build every supported language other than xyz and abc.

ChangeLog:

	PR bootstrap/12407
	* configure.ac: Add feature to parsing of --enable-languages so that
	a language can be disabled by prefixing it with a caret.
	* configure: Regenerate.

gcc/ChangeLog:

	PR bootstrap/12407
	* doc/install.texi (--enable-languages): Document the new language
	exclusion feature.