a68: support only one definition module per prelude packet
At this point we only support having a single definition module in
each prelude packet.  It is planned to change this, and also to
support local not-toplevel modules, but not in the next GCC release.
So until that gets fully implemented, it is pointless to support
parsing multiple contracted module definitions.

This commit removes this support from the parser and adjusts several
tests that were making use of it.

Signed-off-by: Jose E. Marchesi <jemarch@gnu.org>

gcc/algol68/ChangeLog

	* a68-parser-bottom-up.cc (reduce_prelude_packet): Do not support
	multiple module definitions per packet.

gcc/testsuite/ChangeLog

	* algol68/compile/error-compile-unknown-tag-1.a68: Use a single
	module per packet.
	* algol68/compile/error-module-coercions-1.a68: Likewise.
	* algol68/compile/error-module-ranges-1.a68: Likewise.
	* algol68/compile/module-2.a68: Likewise.
	* algol68/compile/module-pub-mangling-1.a68: Likewise.
	* algol68/compile/module-pub-mangling-2.a68: Likewise.
	* algol68/compile/module-pub-mangling-3.a68: Likewise.
	* algol68/compile/module-pub-mangling-4.a68: Likewise.
	* algol68/compile/module-pub-mangling-5.a68: Likewise.
	* algol68/compile/module-pub-mangling-6.a68: Likewise.
	* algol68/compile/warning-module-hidding-1.a68: Likewise.