AVR: Tweak shift execution times in some cases.
When the tail reg (last register) in a shift is an upper register,
then inserting a sequence of 0s and a 1 into the tail register
only takes 2 instruction.  The preparation will be one instruction
longer, but the loop body will be one instruction shorter, saving
count-1 cycles.

For example uint32_t R22 >> 5 will turn from:

    ldi srcatch, 5
1:  lsr r25
    ror r24
    ror r23
    ror r22
    dec scratch
    brne 1b

to:

    andi r22,-32   ; Set lower 5 bits to 0.
    ori r22,16     ; Set bit 4 to 1.
    ;; Now r22 = 0b***10000
1:  lsr r25
    ror r24
    ror r23
    ror r22
    brcc 1b        ; Carry will be 0, 0, 0, 0, 1.

gcc/
	* config/avr/avr.cc (avr_out_shift_with_cnt): Tweak
	execution time by count-1 cycles in some cases.