libstdc++: Move new std::generate_canonical to inline namespace
This ensures that the new definition of std::generate_canonical has a
different mangled name from the old one, so that TUs compiled with GCC
16 will be sure to use the new definition, even if the linker also sees
a symbol instantiated from the old definition. We use the same _V2
inline namespace as used elsewhere (std::_V2::condition_variable,
std::_V2::__rotate, and std::chrono::_V2::system_clock), and use a macro
to add it conditionally so that it's not used for the ABI-unstable
gnu-versioned-namespace configuration.

We can simplify the 26_numerics/random/pr60037-neg.cc test to only use
one dg-error without a line number, so that it matches any of the three
relevant static_assert failures for this test: the one from _Adaptor in
<bits/random.h> and the ones from the new and old definitions of
std::generate_canonical in <bits/random.tcc>. Without this change, the
line number for the dg-error matching the <bits/random.tcc> error
depends on the _GLIBCXX_USE_OLD_GENERATE_CANONICAL macro, which is
awkward to depend on in the test (because DejaGnu sees all dg-error
directives, it doesn't care if they're guarded by #ifdef preprocessor
checks).

libstdc++-v3/ChangeLog:

	* include/bits/random.h [!_GLIBCXX_USE_OLD_GENERATE_CANONICAL]
	(generate_canonical): Use inline namespace _V2.
	* include/bits/random.tcc [!_GLIBCXX_USE_OLD_GENERATE_CANONICAL]
	(generate_canonical): Likewise.
	* testsuite/26_numerics/random/pr60037-neg.cc: Remove lineno so
	that one dg-error matches both diagnostics.

Reviewed-by: Tomasz Kamiński <tkaminsk@redhat.com>