PR modula2/23178: reduce error message clutter when reporting a non procedure
This bug fix corrects the token of an error message when reporting
about a symbol which was not exported from a definition module.
The reimplemented procedure CheckCanBeImported uses MetaError2
with the spell hint format specifier if the symbol is unknown
rather than the older WriteFormat2 procedure.

gcc/m2/ChangeLog:

	PR modula2/23178
	* gm2-compiler/P3SymBuild.mod (CheckCanBeImported): Reimplement to
	use MetaError2 and provide spell hints if possible for unknown symbols.

gcc/testsuite/ChangeLog:

	PR modula2/23178
	* gm2/pim/fail/badprocedure.mod: New test.

Signed-off-by: Gaius Mulley <gaiusmod2@gmail.com>